+++ sdl-deltalake/pom.xml
@@ -112,12 +112,6 @@
 			<artifactId>config</artifactId>
 			<scope>${sdl-core.scope}</scope>
 		</dependency>
-		<dependency>
-			<groupId>com.github.kxbmap</groupId>
-			<artifactId>configs_${scala.minor.version}</artifactId>
-			<scope>${sdl-core.scope}</scope>
-		</dependency>
-
 
 		<!-- TEST dependencies -->
 		<dependency>
@@ -135,6 +129,7 @@
 			<artifactId>scalacheck-1-18_${scala.minor.version}</artifactId>
 			<scope>test</scope>
 		</dependency>
+
 		<!-- sdl-core test utils -->
 		<dependency>
 			<groupId>io.smartdatalake</groupId>
@@ -142,30 +137,6 @@
 			<version>${project.parent.version}</version>
 			<classifier>test-jar</classifier>
 			<type>test-jar</type>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.sshd</groupId>
-			<artifactId>sshd-sftp</artifactId>
-			<version>${sshd.test.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.sshd</groupId>
-			<artifactId>sshd-common</artifactId>
-			<version>${sshd.test.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.sshd</groupId>
-			<artifactId>sshd-core</artifactId>
-			<version>${sshd.test.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.github.tomakehurst</groupId>
-			<artifactId>wiremock-standalone</artifactId>
-			<version>2.25.1</version>
 			<scope>test</scope>
 		</dependency>
